Use dsl options for ksqlDb url
