+++ docs/changes/20250731_progress.md
@@ -106,3 +106,7 @@
 
 ## 2025-07-21 00:46 JST [naruse]
 - confluent.topic.description を削除し Apache Kafka 互換の DLQ 設定に統一
+
+## 2025-07-21 01:35 JST [assistant]
+- ksqlDB client helpers now read from _dslOptions instead of KafkaContextOptions
+- Updated unit tests for new configuration source
