+++ src/Application/KsqlDbResponse.cs
@@ -1,0 +1,3 @@
+namespace Kafka.Ksql.Linq.Application;
+
+public record KsqlDbResponse(bool IsSuccess, string Message, int? ErrorCode = null, string? ErrorDetail = null);
