+++ tests/Application/KsqlDbExecutionExtensionsTests.cs
@@ -4,7 +4,9 @@ using System.Net.Http;
 using System.Reflection;
 using Kafka.Ksql.Linq.Application;
 using Kafka.Ksql.Linq.Core.Context;
-using Microsoft.Extensions.Configuration;
+using Kafka.Ksql.Linq.Configuration;
+using Kafka.Ksql.Linq.Messaging.Configuration;
+using Kafka.Ksql.Linq.Core.Configuration;
 using Xunit;
 
 namespace Kafka.Ksql.Linq.Tests.Application;
@@ -21,17 +23,13 @@ public class KsqlDbExecutionExtensionsTests
     [Fact]
     public void CreateClient_UsesSchemaRegistryHost()
     {
-        var config = new ConfigurationBuilder()
-            .AddInMemoryCollection(new Dictionary<string, string>
-            {
-                { "KsqlDsl:SchemaRegistry:Url", "http://example.com:8085" }
-            })
-            .Build();
+        var ctx = new DummyContext(new KafkaContextOptions());
+        var field = typeof(KsqlContext).GetField("_dslOptions", BindingFlags.NonPublic | BindingFlags.Instance)!;
+        var opts = (KsqlDslOptions)field.GetValue(ctx)!;
+        var schema = opts.SchemaRegistry;
+        typeof(SchemaRegistrySection).GetProperty("Url")!.SetValue(schema, "http://example.com:8085");
 
-        var options = new KafkaContextOptions { Configuration = config };
-        var ctx = new DummyContext(options);
-
-        var method = typeof(KsqlDbExecutionExtensions)
+        var method = typeof(KsqlContext)
             .GetMethod("CreateClient", BindingFlags.NonPublic | BindingFlags.Static)!;
         using var client = (HttpClient)method.Invoke(null, new object[] { ctx })!;
         Assert.Equal(new Uri("http://example.com:8085"), client.BaseAddress);
@@ -40,17 +38,13 @@ public class KsqlDbExecutionExtensionsTests
     [Fact]
     public void CreateClient_FallsBackToBootstrapServers()
     {
-        var config = new ConfigurationBuilder()
-            .AddInMemoryCollection(new Dictionary<string, string>
-            {
-                { "KsqlDsl:Common:BootstrapServers", "example.com:9092" }
-            })
-            .Build();
+        var ctx = new DummyContext(new KafkaContextOptions());
+        var field = typeof(KsqlContext).GetField("_dslOptions", BindingFlags.NonPublic | BindingFlags.Instance)!;
+        var opts = (KsqlDslOptions)field.GetValue(ctx)!;
+        typeof(SchemaRegistrySection).GetProperty("Url")!.SetValue(opts.SchemaRegistry, "");
+        typeof(CommonSection).GetProperty("BootstrapServers")!.SetValue(opts.Common, "example.com:9092");
 
-        var options = new KafkaContextOptions { Configuration = config };
-        var ctx = new DummyContext(options);
-
-        var method = typeof(KsqlDbExecutionExtensions)
+        var method = typeof(KsqlContext)
             .GetMethod("CreateClient", BindingFlags.NonPublic | BindingFlags.Static)!;
         using var client = (HttpClient)method.Invoke(null, new object[] { ctx })!;
         Assert.Equal(new Uri("http://example.com:9092"), client.BaseAddress);
